sql/sem/builtins,sql/catalog/redact: add builtin to redact descriptors
This will be useful by obs infra. The advantage of doing it this way is
that we'll benefit from changes we make to improve redaction over time.

Release note: None
