+++ pkg/BUILD.bazel
@@ -1347,6 +1347,7 @@ GO_TARGETS = [
     "//pkg/sql/catalog/multiregion:multiregion_test",
     "//pkg/sql/catalog/nstree:nstree",
     "//pkg/sql/catalog/nstree:nstree_test",
+    "//pkg/sql/catalog/redact:redact",
     "//pkg/sql/catalog/resolver:resolver",
     "//pkg/sql/catalog/resolver:resolver_test",
     "//pkg/sql/catalog/rewrite:rewrite",
@@ -2543,6 +2544,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/sql/catalog/lease:get_x_data",
     "//pkg/sql/catalog/multiregion:get_x_data",
     "//pkg/sql/catalog/nstree:get_x_data",
+    "//pkg/sql/catalog/redact:get_x_data",
     "//pkg/sql/catalog/resolver:get_x_data",
     "//pkg/sql/catalog/rewrite:get_x_data",
     "//pkg/sql/catalog/schemadesc:get_x_data",
