+++ pkg/ccl/logictestccl/tests/3node-tenant/generated_test.go
@@ -1359,6 +1359,13 @@ func TestTenantLogic_record(
 	runLogicTest(t, "record")
 }
 
+func TestTenantLogic_redact_descriptor(
+	t *testing.T,
+) {
+	defer leaktest.AfterTest(t)()
+	runLogicTest(t, "redact_descriptor")
+}
+
 func TestTenantLogic_rename_atomic(
 	t *testing.T,
 ) {
@@ -2148,6 +2155,13 @@ func TestTenantLogicCCL_partitioning_enum(
 ) {
 	defer leaktest.AfterTest(t)()
 	runCCLLogicTest(t, "partitioning_enum")
+}
+
+func TestTenantLogicCCL_redact_descriptor(
+	t *testing.T,
+) {
+	defer leaktest.AfterTest(t)()
+	runCCLLogicTest(t, "redact_descriptor")
 }
 
 func TestTenantLogicCCL_schema_change_in_txn(
