+++ pkg/ccl/logictestccl/tests/fakedist/BUILD.bazel
@@ -10,7 +10,7 @@ go_test(
         "//c-deps:libgeos",  # keep
         "//pkg/ccl/logictestccl:testdata",  # keep
     ],
-    shard_count = 4,
+    shard_count = 5,
     tags = ["cpu:2"],
     deps = [
         "//pkg/build/bazel",
