+++ pkg/ccl/logictestccl/tests/local-vec-off/BUILD.bazel
@@ -10,7 +10,7 @@ go_test(
         "//c-deps:libgeos",  # keep
         "//pkg/ccl/logictestccl:testdata",  # keep
     ],
-    shard_count = 3,
+    shard_count = 4,
     tags = ["cpu:1"],
     deps = [
         "//pkg/build/bazel",
