+++ pkg/sql/catalog/schematelemetry/BUILD.bazel
@@ -20,15 +20,11 @@ go_library(
         "//pkg/settings/cluster",
         "//pkg/sql",
         "//pkg/sql/catalog",
-        "//pkg/sql/catalog/catpb",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/nstree",
+        "//pkg/sql/catalog/redact",
         "//pkg/sql/catalog/schematelemetry/schematelemetrycontroller",
-        "//pkg/sql/catalog/tabledesc",
-        "//pkg/sql/catalog/typedesc",
-        "//pkg/sql/parser",
-        "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/builtins/builtinconstants",
         "//pkg/sql/sem/tree",
