+++ pkg/sql/evalcatalog/BUILD.bazel
@@ -17,8 +17,10 @@ go_library(
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/sql/catalog",
+        "//pkg/sql/catalog/descbuilder",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
+        "//pkg/sql/catalog/redact",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/rowenc",
@@ -26,6 +28,8 @@ go_library(
         "//pkg/sql/sem/tree",
         "//pkg/sql/sqlerrors",
         "//pkg/sql/types",
+        "//pkg/util/hlc",
+        "//pkg/util/protoutil",
         "@com_github_cockroachdb_errors//:errors",
     ],
 )
