+++ pkg/sql/logictest/tests/local/generated_test.go
@@ -1451,6 +1451,13 @@ func TestLogic_record(
 	runLogicTest(t, "record")
 }
 
+func TestLogic_redact_descriptor(
+	t *testing.T,
+) {
+	defer leaktest.AfterTest(t)()
+	runLogicTest(t, "redact_descriptor")
+}
+
 func TestLogic_rename_atomic(
 	t *testing.T,
 ) {
